make token list a separate screen inside tokens list modal, so adding custom token may be implemented
