+++ src/routes/safe/component/Balances/AssetTableCell/AssetTableCell.js
@@ -3,7 +3,7 @@ import * as React from 'react'
 import Block from '~/components/layout/Block'
 import Img from '~/components/layout/Img'
 import Paragraph from '~/components/layout/Paragraph'
-import TokenPlaceholder from '../Tokens/assets/token_placeholder.svg'
+import TokenPlaceholder from '../assets/token_placeholder.svg'
 
 const setImageToPlaceholder = (e) => {
   e.target.onerror = null
