--- src/routes/safe/component/Balances/Tokens/screens/tokenList/TokenList.js
@@ -1,13 +1,0 @@
-// @flow
-
-import React, { Component } from 'react'
-
-export default class TokenLIst extends Component {
-  render() {
-    return (
-      <div>
-        token list
-      </div>
-    )
-  }
-}
