--- src/routes/safe/component/Balances/Tokens/screens/tokenList/index.js
@@ -1,4 +1,0 @@
-// @flow
-import TokenList from './TokenList'
-
-export default TokenList
