+++ src/routes/safe/component/Balances/utils/index.js
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './setTokenImgToPlaceholder'
